+++ CHANGELOG.md
@@ -20,7 +20,8 @@ Changes are grouped as follows
 ## [6.2.0] - 2023-05-24
 ### Added
 - Support for the spaces resource in the Data Modeling API with the implementation `client.data_modeling.spaces`.
-
+### Improved
+- Reorganized documentation to match API documentation.
 
 ## [6.1.10] - 2023-05-22
 ### Fixed